ADD: homeworks/7-timer-notifications
