+++ homeworks/7-timer-notifications/README.md
@@ -1,0 +1,17 @@
+# Таймер
+
+Время (час, минута, секунда) указывается в аргументах команды node в любом порядке:
+
+* Час - аргумент содержит букву `h`, например: 1h
+* Минута - аргумент содержит букву `m`, например: 23m
+* Секунда - аргумент содержит букву `s`, например: 10m
+* Задача - аргумент содержит подстроку `task=`, после знака равно, в двойных кавычках указывается задача, например: task="Перерыв!". Если задачу не указать, для задачи будет использоваться текст: "Таймер завершен!"
+
+Примеры:
+node app.js 1h 10m 5s task="Позвонить другу";
+node app.js task="Сделать перерыв" 15s 2m 2h;
+node app.js 33s;
+node app.js 30m 1h;
+и т.д.
+
+По завершению таймера будет отправлено уведомление с текстом задачи
