+++ homeworks/7-timer-notifications/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "timer-notifications",
+  "version": "1.0.0",
+  "description": "",
+  "main": "app.js",
+  "keywords": [],
+  "author": "Artyom Ugryumov",
+  "license": "ISC",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "dependencies": {
+    "node-notifier": "^10.0.1"
+  }
+}
